+++ trunk/cpsycle/audio/src/preset.h
@@ -13,7 +13,7 @@ extern "C" {
 typedef struct psy_audio_Preset {
     char* name;
     psy_Table parameters;
-    int32_t datasize;
+    uintptr_t datasize;
     unsigned char* data;
     bool isfloat;
     int32_t magic;
@@ -31,8 +31,8 @@ const char* psy_audio_preset_name(psy_audio_Preset*);
 void psy_audio_preset_setvalue(psy_audio_Preset*, uintptr_t param, intptr_t value);
 intptr_t psy_audio_preset_value(psy_audio_Preset*, uintptr_t param);
 void psy_audio_preset_setdatastruct(psy_audio_Preset* self,
-    int num, const char* newname, int const* parameters, int size, void* newdata);
-void psy_audio_preset_putdata(psy_audio_Preset*, int size, void* newdata);
+    uintptr_t num, const char* newname, int const* parameters, uintptr_t size, void* newdata);
+void psy_audio_preset_putdata(psy_audio_Preset*, uintptr_t size, void* newdata);
 
 INLINE uintptr_t psy_audio_preset_numparameters(psy_audio_Preset* self)
 {
